+++ src/libfsm/internal.h
@@ -13,11 +13,10 @@
 #include <fsm/fsm.h>
 #include <fsm/options.h>
 
-#include <adt/set.h>
+struct edge_set;
+struct state_set;
 
 #define FSM_ENDCOUNT_MAX ULONG_MAX
-
-/* struct set; */
 
 /*
  * The alphabet (Sigma) for libfsm's FSM is arbitrary octets.
@@ -36,95 +35,6 @@ enum fsm_edge_type {
 };
 
 #define FSM_EDGE_MAX FSM_EDGE_EPSILON
-
-struct fsm_edge;
-struct fsm_state;
-
-struct edge_set {
-	struct set *set;
-};
-
-struct edge_iter {
-	struct set_iter iter;
-};
-
-void
-edge_set_free(struct edge_set *set);
-
-struct fsm_edge *
-edge_set_add(struct edge_set *set, struct fsm_edge *e);
-
-struct fsm_edge *
-edge_set_contains(const struct edge_set *set, const struct fsm_edge *e);
-
-size_t
-edge_set_count(const struct edge_set *set);
-
-void
-edge_set_remove(struct edge_set *set, const struct fsm_edge *e);
-
-struct fsm_edge *
-edge_set_first(struct edge_set *set, struct edge_iter *it);
-
-struct fsm_edge *
-edge_set_firstafter(struct edge_set *set, struct edge_iter *it, const struct fsm_edge *e);
-
-struct fsm_edge *
-edge_set_next(struct edge_iter *it);
-
-int
-edge_set_hasnext(struct edge_iter *it);
-
-struct fsm_edge *
-edge_set_only(const struct edge_set *s);
-
-struct state_set {
-	struct set *set;
-};
-
-struct state_set *
-state_set_create(void);
-
-void
-state_set_free(struct state_set *set);
-
-struct fsm_state *
-state_set_add(struct state_set *set, struct fsm_state *st);
-
-void
-state_set_remove(struct state_set *set, const struct fsm_state *st);
-
-int
-state_set_empty(const struct state_set *s);
-
-struct fsm_state *
-state_set_only(const struct state_set *s);
-
-struct state_iter {
-	struct set_iter iter;
-};
-
-struct fsm_state *
-state_set_contains(const struct state_set *set, const struct fsm_state *st);
-
-size_t
-state_set_count(const struct state_set *set);
-
-struct fsm_state *
-state_set_first(struct state_set *set, struct state_iter *it);
-
-struct fsm_state *
-state_set_next(struct state_iter *it);
-
-int
-state_set_hasnext(struct state_iter *it);
-
-/* uses set_cmp */
-int
-state_set_cmp(const struct state_set *a, const struct state_set *b);
-
-const struct fsm_state **
-state_set_array(const struct state_set *set);
 
 struct fsm_edge {
 	struct state_set *sl;
