+++ src/libfsm/mergestates.c
@@ -12,6 +12,8 @@
 #include <fsm/fsm.h>
 
 #include "internal.h"
+#include "stateset.h"
+#include "edgeset.h"
 
 struct fsm_state *
 fsm_mergestates(struct fsm *fsm, struct fsm_state *a, struct fsm_state *b)
