+++ src/libfsm/pred/epsilonsonly.c
@@ -12,6 +12,8 @@
 #include <fsm/pred.h>
 
 #include "../internal.h"
+#include "../stateset.h"
+#include "../edgeset.h"
 
 int
 fsm_epsilonsonly(const struct fsm *fsm, const struct fsm_state *state)
