+++ src/libfsm/pred/iscomplete.c
@@ -13,6 +13,8 @@
 #include <fsm/pred.h>
 
 #include "../internal.h"
+#include "../stateset.h"
+#include "../edgeset.h"
 
 int
 fsm_iscomplete(const struct fsm *fsm, const struct fsm_state *state)
