+++ src/libfsm/print/fsm.c
@@ -20,6 +20,9 @@
 #include <fsm/print.h>
 #include <fsm/options.h>
 
+#include "libfsm/stateset.h"
+#include "libfsm/edgeset.h"
+
 static unsigned int
 indexof(const struct fsm *fsm, const struct fsm_state *state)
 {
