+++ src/libfsm/print/json.c
@@ -17,6 +17,8 @@
 #include <fsm/options.h>
 
 #include "libfsm/internal.h"
+#include "libfsm/stateset.h"
+#include "libfsm/edgeset.h"
 
 static unsigned int
 indexof(const struct fsm *fsm, const struct fsm_state *state)
