+++ src/libfsm/state.c
@@ -12,6 +12,8 @@
 #include <fsm/fsm.h>
 
 #include "internal.h"
+#include "stateset.h"
+#include "edgeset.h"
 
 static int
 fsm_state_cmpedges(const void *a, const void *b)
