+++ src/libfsm/walk/reachable.c
@@ -13,6 +13,8 @@
 #include <fsm/fsm.h>
 
 #include "../internal.h"
+#include "../stateset.h"
+#include "../edgeset.h"
 
 static int
 fsm_reachable(const struct fsm *fsm, const struct fsm_state *state,
